98% completion...
The program seems to run with no issues/crashes. Need some debugging to be done by someone first before being determined as 100% complete. Two Java files exist in repo; the one in "debug_menu" has debugging functionality, whereas the default package is the one to be submitted to Mr. So. Use the "debug_menu" file to test program. When debugging is complete, let me know.